modified rsync.sh

+++ rsync.sh
@@ -4,22 +4,31 @@ if [[ $# -eq 0 ]] ; then
     exit 0
 fi
 
-while [[ $# -gt 0 ]]
-do
+#while [[ $# -gt 0 ]]
+#do
 key="$1"
 case $key in
     pull)
-        rsync -azvP 'userver@192.157.226.233':~/anywhere  ~/repositories/anywhere/  
+        if [ "$2" == 'remote' ]
+        then
+            rsync -azvP 'userver@192.157.226.233':~/anywhere  ~/repositories/anywhere/  
+        elif [ "$2" == 'local' ]
+        then
+            rsync -azvP --delete /run/media/xFrog/backup/anywhere/elts/  ~/Documents/elts 
+        fi
         ;;
-    usb)
-        rsync -azvP --delete ~/Downloads/anywhere/elts/ /run/media/zhouyu/backup/anywhere/elts  
-        ;;
-    s)
-        rsync -azvP --delete ~/Downloads/anywhere/elts/ 'userver@192.157.226.233':~/anywhere/elts
+    push)
+        if [ "$2" == 'local' ]
+        then
+            rsync -azvP --delete ~/Downloads/anywhere/elts/ /run/media/zhouyu/backup/anywhere/elts  
+        elif [ "$2" == 'remote' ]
+        then
+            rsync -azvP --delete ~/Downloads/anywhere/elts/ 'userver@192.157.226.233':~/anywhere/elts
+        fi
         ;;
     *)
         ;;
 esac
-shift #past argument or value
-done
+#shift #past argument or value
+#done
